# Refactor the scanning step of the parser