… Sources/WebURL/Util/BidirectionalCollection+suffix.swift
… Sources/WebURL/Util/FastCollectionAlgorithms.swift
… Sources/WebURL/Util/MutableCollection+pathUtils.swift
… Sources/WebURL/WebURL+FilePaths.swift
… Sources/WebURL/WebURL+FormParameters.swift
… Sources/WebURL/WebURL+JSModel.swift
… Sources/WebURL/WebURL+PathComponents.swift
… Tests/WebURLTests/CollectionExtensionsTests.swift