Some pretty big changes:

1. Constructs empty mappings and populates their data in-place. These are pretty big values, so we don't want to be passing them around as parameters or return values. This saves a lot of binary size which was spent just shuffling data between registers.
2. Switches (heh) from a state machine to just tail-calling the next scanning step. Apparently the compiler is able to reason about the control flow better, so we see more inlining, loop merging, etc. It's pretty cool, actually.
3. Moves from static methods to mutating instance methods.
4. Removed some helper types which are no longer needed. Lots of binary size savings!
5. Better comments and documentation (I think). It was already quite good, but now it's better.

Binary size is down by 2.4%. Which might not seem like a huge amount, but isn't bad, either. They won't all be whopping 20% reductions.

```
Title                              Section             Old             New  Percent
WebURLBenchmark                     __text:        1418369         1383825    -2.4%
```

Performance is better, too.

x86 Mac:

```
name                                                   time         std        iterations
-----------------------------------------------------------------------------------------
Constructor.SpecialNonFile.AverageURLs                 20008.000 ns ±  14.81 %      60797
Constructor.SpecialNonFile.AverageURLs filtered        37435.000 ns ±  16.95 %      35084
Constructor.SpecialNonFile.IPv4 host                   18377.000 ns ±  16.86 %      64460
Constructor.SpecialNonFile.IPv4 host filtered          28475.000 ns ±  13.34 %      44093
Constructor.SpecialNonFile.IPv6 host                   23199.000 ns ±  16.81 %      56333
Constructor.SpecialNonFile.IPv6 host filtered          27529.000 ns ±  15.53 %      47831
Constructor.SpecialNonFile.Percent-encoding components  9344.000 ns ±  14.61 %     139110
Constructor.SpecialNonFile.Percent-encoded hostnames    8599.000 ns ±  16.06 %     137821
Constructor.SpecialNonFile.Long paths                  74211.000 ns ±  10.41 %      18109
Constructor.SpecialNonFile.Complex paths 1             23397.000 ns ±  12.34 %      55674
Constructor.SpecialNonFile.Complex paths 2             27512.000 ns ±  16.52 %      47446
Constructor.SpecialNonFile.Long query 1                 2064.000 ns ±  23.22 %     629974
Constructor.SpecialNonFile.Long query 2                20668.000 ns ±  11.41 %      61285
```

We're juuuust above the 20ms mark for AverageURLs, which was around 22ms before. IPv4 host benchmarks see an even bigger benefit (those benchmarks don't include paths or queries, which is why they're so much lower than AverageURL -- they're just smaller strings). Long paths also see a pretty nice reduction, from about 80ms to 74ms.

On the Raspberry Pi, there's a bigger improvement, from 53ms to 46ms (-13%).

Before:

```
name                                                   time           std        iterations warmup          
------------------------------------------------------------------------------------------------------------
Constructor.SpecialNonFile.AverageURLs                   53796.000 ns ±   2.32 %      25836  11408117.000 ns
Constructor.SpecialNonFile.AverageURLs filtered          89740.000 ns ±   1.86 %      15565  18122229.000 ns
Constructor.SpecialNonFile.IPv4 host                     63259.000 ns ±   2.16 %      22079  12791404.000 ns
Constructor.SpecialNonFile.IPv4 host filtered            85185.000 ns ±   1.88 %      16389  17281483.000 ns
Constructor.SpecialNonFile.IPv6 host                     60592.000 ns ±   3.67 %      23130  12293682.000 ns
Constructor.SpecialNonFile.IPv6 host filtered            74481.000 ns ±   2.90 %      18776  15072683.000 ns
Constructor.SpecialNonFile.Percent-encoding components   24074.000 ns ±   3.25 %      57979   4883374.000 ns
Constructor.SpecialNonFile.Percent-encoded hostnames     25296.000 ns ±   6.15 %      55164   5208342.000 ns
Constructor.SpecialNonFile.Long paths                   173240.000 ns ±   1.77 %       8062  34893784.000 ns
Constructor.SpecialNonFile.Complex paths 1               56833.000 ns ±   1.94 %      24583  11428409.000 ns
Constructor.SpecialNonFile.Complex paths 2               70444.000 ns ±   1.74 %      19834  14274617.000 ns
Constructor.SpecialNonFile.Long query 1                   5796.000 ns ±   6.93 %     240968   1170608.000 ns
Constructor.SpecialNonFile.Long query 2                  78758.000 ns ±   2.37 %      17795  16047965.000 ns
```

After:

```
name                                                   time           std        iterations warmup          
------------------------------------------------------------------------------------------------------------
Constructor.SpecialNonFile.AverageURLs                   46889.000 ns ±   2.36 %      29688  10082653.000 ns
Constructor.SpecialNonFile.AverageURLs filtered          82407.000 ns ±   1.77 %      16944  16687864.000 ns
Constructor.SpecialNonFile.IPv4 host                     49296.000 ns ±   4.01 %      28187  10108000.000 ns
Constructor.SpecialNonFile.IPv4 host filtered            78851.000 ns ±   2.08 %      17760  15844325.000 ns
Constructor.SpecialNonFile.IPv6 host                     53592.000 ns ±   2.44 %      26036  10881132.000 ns
Constructor.SpecialNonFile.IPv6 host filtered            65555.000 ns ±   3.96 %      21158  13309628.000 ns
Constructor.SpecialNonFile.Percent-encoding components   22259.000 ns ±   4.11 %      62799   4547095.000 ns
Constructor.SpecialNonFile.Percent-encoded hostnames     23278.000 ns ±   4.88 %      59849   4807787.000 ns
Constructor.SpecialNonFile.Long paths                   162091.000 ns ±   1.91 %       8531  32656950.000 ns
Constructor.SpecialNonFile.Complex paths 1               52241.000 ns ±   2.63 %      26525  10559268.000 ns
Constructor.SpecialNonFile.Complex paths 2               64740.000 ns ±   2.16 %      21563  13082609.000 ns
Constructor.SpecialNonFile.Long query 1                   4611.000 ns ±   9.12 %     303351    935619.000 ns
Constructor.SpecialNonFile.Long query 2                  69833.000 ns ±   2.26 %      19917  14347742.000 ns
```

For some reason, IP addresses are slower than AverageURLs on the Pi. Not quite sure why, but it was like that before these changes, too. I need to have a look at what the IP address parser is doing on these machines. It could be lookup-table related (for the ASCII -> number parsing), but that's just a guess.